% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_SP.R
\docType{data}
\name{dengue_SP}
\alias{dengue_SP}
\title{Dengue cases from the "S達o Paulo" state of Brazil}
\format{
A data frame with 678 rows and 8 columns:
\describe{
\item{\code{date}}{First day of the week, in date format ("\%d-\%m-\%Y")}
\item{\code{geocode}}{Unique ID code for S達o Paulo microregion}
\item{\code{cases}}{Number of notified dengue cases}
\item{\code{year}}{Year 2000 - 2022}
\item{\code{temp_med}}{Weekly average daily mean temperature}
\item{\code{precip_tot}}{Weekly cumulative precipitation}
\item{\code{enso}}{El Ni単o-Southern Oscillation Index}
\item{\code{pop}}{Number of inhabitants}
}
}
\source{
\href{https://info.dengue.mat.br/services/api}{Infodengue API}
}
\usage{
dengue_SP
}
\description{
The \code{dengue_SP} example data set reports the weekly number of notified dengue cases in
the municipality of S達o Paulo together with climatic covariates. Data was sourced from
Infodengue (see \code{@source}).
}
\keyword{datasets}
