% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHRmodel-package.R
\docType{package}
\name{GHRmodel-package}
\alias{GHRmodel}
\alias{GHRmodel-package}
\title{GHRmodel: Bayesian Hierarchical Modelling of Spatio-Temporal Health Data}
\description{
Supports modeling health outcomes using Bayesian hierarchical spatio-temporal models with complex covariate effects (e.g., linear, non-linear, interactions, distributed lag linear and non-linear models) in the 'INLA' framework. It is designed to help users identify key drivers and predictors of disease risk by enabling streamlined model exploration, comparison, and visualization of complex covariate effects. See an application of the modelling framework in Lowe, Lee, O'Reilly et al (2021) \doi{10.1016/S2542-5196(20)30292-8}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.earth.bsc.es/ghr/ghrmodel}
  \item \url{https://bsc-es.github.io/GHRtools/docs/GHRmodel/GHRmodel}
  \item Report bugs at \url{https://gitlab.earth.bsc.es/ghr/ghrmodel/-/issues}
}

}
\author{
\strong{Maintainer}: Carles Milà \email{carles.milagarcia@bsc.es} (\href{https://orcid.org/0000-0003-0470-0760}{ORCID})

Authors:
\itemize{
  \item Giovenale Moirano (\href{https://orcid.org/0000-0001-8748-3321}{ORCID})
  \item Anna B. Kawiecki (\href{https://orcid.org/0000-0002-0499-2612}{ORCID})
  \item Rachel Lowe (\href{https://orcid.org/0000-0003-3939-7343}{ORCID})
}

}
\keyword{internal}
